fix: emoji-data script and simplify emoji logic
